Minor changes to interpreter

+++ apps/hardware_benchmarks/hw_support/coreir_interpret.cpp
@@ -486,12 +486,13 @@ void run_coreir_on_interpreter(string coreir_design,
           if (valid_value) {
             T output_value = state.getBitVec(output_name).to_type<T>();
             coreir_img_writer.write(output_value);
-            std::cout << "y=" << y << ",x=" << x << " " << hex << "in=" << (input(x,y,c) & 0xff) << " out=" << output_value << dec << endl;
+            std::cout << "y=" << y << ",x=" << x << " " << hex << "in=" << (((unsigned long long)input(x,y,c)) & 0xff) << " out=" << output_value << dec << endl;
           }
         } else {
           T output_value = state.getBitVec(output_name).to_type<T>();
           output(x,y,c) = output_value;
-          std::cout << "y=" << y << ",x=" << x << " " << hex << "in=" << (input(x,y,c) & 0xff) << " out=" << output_value << dec << endl;
+          std::cout << "y=" << y << ",x=" << x << " " << hex << "in=" << (((unsigned long long)input(x,y,c)) & 0xff) << " out=" << output_value << dec << endl;
+          //std::cout << "y=" << y << ",x=" << x << " " << hex << "in=" << (input(x,y,c) & 0xff) << " out=" << output_value << dec << endl;
         }
